Create Tower of Hanoi
Tower of Hanoi Problem
The Tower of Hanoi is a classic problem in computer science and mathematics that involves moving a stack of disks from one rod to another, following specific rules. 

Problem Description

Given three rods and n disks of different sizes, the objective is to move all the disks from the source rod to the destination rod, following these rules:

-Only one disk can be moved at a time.
-A disk can only be moved if it is the uppermost disk on a rod.
-No disk may be placed on top of a smaller disk.
The rods are typically labeled as Source (src), Auxiliary (aux), and Destination (dest).

Approach to Solve the Problem

The Tower of Hanoi problem can be efficiently solved using a recursive algorithm. 
Base Case:
 If there is only one disk to move (n==1), simply move it from the source rod to the destination rod. 

Recursive Case: 
Move the top n−1 disks from the source rod to the auxiliary rod, using the destination rod as a temporary holding area.
Move the remaining largest disk from the source rod to the destination rod. Finally, move the n−1 disks from the auxiliary rod to the destination rod, using the source rod as a temporary holding area.