Create Binary Search using Recursion
Binary Search

Binary search leverages the "divide and conquer" strategy to systematically narrow down the search space/

Core Principles:

Sorted Input: Binary search assumes the input array is sorted in ascending order. This is crucial for its efficient operation. If the array is unsorted, the results will be unpredictable.

Repeated Division: The algorithm repeatedly divides the search space (the portion of the array that might potentially contain the target) in half.

Middle Element Comparison: At each iteration, it compares the target value with the element at the middle index of the current search space.

Recursive Branching: 
If target equals the value of the element at index same as mid then the target is found, and the function returns the index of the element.
If target is less than the value of the element at index same as mid, the target can only exist in the left half of the search space. The recursive call continues searching in left half.
If target is greater than the value of the element at index same as mid, the target can only exist in the right half of the search space. The recursive call continues searching in that half.