Create Unique Birthday Gift
Question: He selected two numbers, N and K and decided to write down on paper all integer arrays of length K
(in form a[1], a[2], ..., a[K]), where every number a[i] is in range from 1 to N, and, moreover, a[i+1] is
divisible by a[i] (where 1 < i <= K), and give you this paper as a birthday present. Since the answer can be really large, print it modulo 10000. 
Input:
The first line contains an integer, n, denoting the maximum possible value in the arrays.
The next line contains an integer, k, denoting the length of the arrays.