Create Design a stack that supports getMin() in O(1) time and O(1) extra space
Approach

Push: If the element is greater than min_element then push it. If not then push (2*element-mini) and update mini.

Pop: If element is greater than min_element then pop. If not update mini as (2*mini-element) and then pop.

Top: If element is greater than min_element then return top of stack. If not return mini.